sql/parser: types.go -> new package sem/types
