+++ pkg/sql/parser/expr.go
@@ -17,6 +17,8 @@ package parser
 import (
 	"bytes"
 	"fmt"
+
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/types"
 )
 
 // Expr represents an expression.
@@ -1123,7 +1125,7 @@ var (
 
 // validCastTypes returns a set of types that can be cast into the provided type.
 func validCastTypes(t Type) []Type {
-	switch UnwrapType(t) {
+	switch types.UnwrapType(t) {
 	case TypeBool:
 		return boolCastTypes
 	case TypeInt:
